lots of bug fixes.  Things look pretty good now.
git-svn-id: https://svn.atnf.csiro.au/difx@564 9d4a3555-dc7e-4673-bb7d-d8a0b9a30876

+++ utilities/mk5daemon/trunk/src/logger.c
@@ -72,7 +72,7 @@ int Logger_logData(Logger *log, const char *message)
 	
 	if(t != log->lastTime)
 	{
-		mjd = (int)(40587.0 + t/86400.0);
+		mjd = 40587.0 + t/86400.0;
 
 		if(t/86400 != log->lastTime/86400)
 		{
